Fix inconsistent indexing between R and C++
The previous commit accidentally introduced an error: it created
a redundant correction for indices in the C++, although the indices
were already corrected in the R (i.e., -1 to start from 0). This
commit removes the correction on the R side, as it seems clearer to
do it in the C++ file.