Fill in gaps in covariates before creating model matrices
So far, gaps in covariates were filled (replaced by nearest available
value) in HMM$initialize, *after* the MarkovChain and Observation
components of the model were created. As a consequence, the data
going into the creation of model terms for the MarkovChain and
Observation objects (model matrices...) still included the missing
values. By default, mgcv excludes missing values, and so the resulting
model matrices had the wrong number of rows. The gaps are now filled
before creating the model matrices.