Fix HMM$post_coeff when fixed parameters
In a model with fixed parameters (either by definition or by
user specification), posterior samples only include those
parameters that are estimated, leading to errors (when trying
to multiply them by the design matrices for example). To solve
this, I created a matrix par_array, which records the values
of all parameters and whether they are fixed. This is created
at setup, and then updated after model fitting.