remove custom distribution functions
Some code was meant to allow the user to define a new observation
distribution. This did not work correctly, and caused problems
with building the package (e.g., required temporarily copying
compilation files to the right path). This was also not written
to work across platforms, and was only tested on Linux. This
commit removes that code, to improve cross-platform compatibility.
User-defined distributions might be implemented at a later stage.