allow for ID-specific delta0
The initial distribution delta0 was so far assumed to be the same for
all time series in the data, but there is no reason why this would
be the case in general. delta0 is now stored as a matrix with one row
for each ID and one column for each state (i.e., row i is the initial
distribution for ith time series). This commit might break code to
set constraints on delta0, and this will need to be tested and fixed
separately.