transform delta0 even when constraints
delta0 was not transformed when it was included in fixpar, to avoid
numerical problems if one of its elements was fixed to 0 (which seems
plausible). This caused other problems, i.e., the entries of delta0
were unbounded and sometimes estimated as < 0 or > 1. The entries are
now always transformed to make sure they sum to 1. If one element needs
to be fixed at zero, a very small positive number might need to be used
instead.