allow for user-defined reference transition probabities
The transition probabilities used as references for the rows (i.e., the
element of each row of the tpm that is not estimated) were assumed to be
along the diagonal. Now the user can pass a vector of indices to specify
the references when creating the MarkovChain object. This makes it
possible to implement complex models, e.g., where a diagonal transition
probability needs to be fixed to zero.