add init = "par" option to tmbstan
This argument means that the sampler starts from the initial parameters
provided to TMB (rather than the default, which is for Stan to generate
random starting values). This makes it impossible for users to
specify the init argument themselves, which might be a problem in some
cases; if so, small changes will need to be made.