minor fix to allow for tibble input data
Data formatted as a tibble caused an error in some functions, because
something like data[1, "myvar"] returns a tibble when data is a
tibble (whereas it returns a numeric when data is a dataframe). I've
changed the syntax to data[[myvar]][i], which works the same for
dataframes and tibbles.