use ID-specific delta0 when initial_state = "stationary"
So far, when the initial state was set to the stationary distribution,
the tpm for the first observation was used for all IDs. This commit
fixes this problem by setting delta0 to something different for
each ID. I'm not sure why the whole C++ file is marked as changed,
probably something to do with auto-indent.