replace optimx by nlminb
Optimisation was done using optimx for flexibility (e.g., multiple
choices of optimisers), but this also caused some issues. In
particular, it seems that there was no way to specify the control
options eval.max and iter.max separately in optimx, which caused
the optimisation to often stop prematurely for complex models.