Use generalized determinant for penalty matrices
The (log)determinant of penalty matrices is used in the
log-likelihood, and was so far computed directly in C++
every time the function is evaluated. It is now computed
only once, on the R side, in make_matrices(). The updated
code also uses the generalized determinant which works for
rank deficient penalty matrices. Thanks @janoleko for your
help with this!