automatically re-order fixed parameters if needed
Automatically orders vector of fixed parameters (e.g.,
NAs and integers) to be in same order as coeff_fe by
matching parameter names. This came up in Discussion #34
because hmmTMB automatically fixes the size of the binomial
distribution, and those constraints were added to user-
defined constraints on the probability parameter of the
binomial (i.e., fixpar was ordered as prob first and
size after, whereas coeff_fe is the opposite).