add error message for laplace = TRUE
Passing laplace = TRUE to fit_stan() isn't currently
working because it doesn't properly deal with the number
of fixed vs random parameters. To make it work I wonder
if the 'init' argument would have to be set to "random"
in that case, and parameter unpacking would have to be
changed when 'iters' is filled. But laplace = TRUE doesn't
seem to speed things up on simple examples, and isn't
recommended by the tmbstan documentation, so probably best
to avoid anyway.